+++ lib/hermes/tracers.rb
@@ -1,0 +1,5 @@
+module Hermes
+  module Tracers
+    autoload :Datadog, "hermes/tracers/datadog"
+  end
+end